@@ -2,10 +2,10 @@ package org.ea.sqrl.activites.base;
 
 import android.content.res.Configuration;
 import android.os.Bundle;
-import android.support.annotation.StringRes;
-import android.support.v7.app.AlertDialog;
-import android.support.v7.app.AppCompatActivity;
-import android.support.v7.app.AppCompatDelegate;
+
+import androidx.appcompat.app.AlertDialog;
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatDelegate;
 
 import org.ea.sqrl.R;
 import org.ea.sqrl.utils.Utils;
@@ -77,7 +77,7 @@ public class CommonBaseActivity extends AppCompatActivity {
                 .show();
     }
 
-    protected void showInfoMessage(@StringRes int title, @StringRes int message) {
+    protected void showInfoMessage(int title, int message) {
         showInfoMessage(
                 this.getResources().getString(title),
                 this.getResources().getString(message),
@@ -85,7 +85,7 @@ public class CommonBaseActivity extends AppCompatActivity {
         );
     }
 
-    protected void showInfoMessage(@StringRes int title, @StringRes int message, Runnable done) {
+    protected void showInfoMessage(int title, int message, Runnable done) {
         showInfoMessage(
                 this.getResources().getString(title),
                 this.getResources().getString(message),
